+++ crates/api/src/measurement/num_traits/cast.rs
@@ -1,4 +1,4 @@
-use num_traits::ToPrimitive;
+use num_traits::{FromPrimitive, ToPrimitive};
 
 use crate::Measurement;
 
@@ -17,6 +17,20 @@ impl ToPrimitive for Measurement {
 
     fn to_f64(&self) -> Option<f64> {
         Some(f64::from(self))
+    }
+}
+
+impl FromPrimitive for Measurement {
+    fn from_i64(n: i64) -> Option<Self> {
+        n.to_f64().and_then(FromPrimitive::from_f64)
+    }
+
+    fn from_u64(n: u64) -> Option<Self> {
+        n.to_f64().and_then(FromPrimitive::from_f64)
+    }
+
+    fn from_f64(n: f64) -> Option<Self> {
+        <Self as num_traits::NumCast>::from(n)
     }
 }
 
@@ -56,6 +70,21 @@ mod tests {
         assert_ulps_eq!(-0.01, f64::from(&m));
         assert_eq!(0, m.to_i64().unwrap());
         assert_eq!(0, m.to_u64().unwrap());
+    }
+
+    #[test]
+    fn from_primitive_test() {
+        let output = Measurement::from_u32(std::u32::MAX).unwrap();
+        assert_eq!(
+            Measurement::new(std::u32::MAX.into(), crate::unit::UNITY),
+            output
+        );
+
+        let output = Measurement::from_i32(std::i32::MAX).unwrap();
+        assert_eq!(
+            Measurement::new(std::i32::MAX.into(), crate::unit::UNITY),
+            output
+        );
     }
 
     #[test]